requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
mysql-connector-python==8.2.0
PyJWT==2.12.1
aiofiles==23.2.1
email-validator==2.1.0
pytz==2023.3